+++ packages/tools/CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Upcoming...
 
 - ... <!-- Add new lines here. -->
+- feat: Move `combine`-/`extractAppendixesAndComments` and `eliminateComments`
+  to `utils` – deprecate importing them from `cleanupEditorOutputs`
 - feat: Add type `RegulationTextProps`
 - fix(ts): Import `globals.d.ts` where `htmldiff-js` typings are required –
   (helps local monorepo builds run)